+++ scalyr_agent/config_main.py
@@ -37,6 +37,11 @@ from distutils import spawn
 from pwd import getpwnam
 from optparse import OptionParser
 
+from scalyr_agent.platform_controller import PlatformController, TARBALL_INSTALL
+
+# By importing the platform modules, they register themselves if they apply to the current platform.
+# noinspection PyUnresolvedReferences
+import scalyr_agent.platform_posix
 
 from __scalyr__ import scalyr_init, determine_file_path
 
@@ -194,19 +199,22 @@ def upgrade_tarball_install(config, new_tarball, preserve_old_install):
 
     try:
         try:
+            platform_controller = PlatformController.new_platform()
+            default_paths = platform_controller.default_paths
+
             # Ensure that this is a tarball install
-            if not config.is_tarball_install():
+            if not platform_controller.install_type != TARBALL_INSTALL:
                 raise UpgradeFailure('The current agent was not installed using a tarball, so you may not use the '
                                      'upgrade tarball command.')
 
             # Ensure that the user has not changed the defaults for the config, data, and log directory.
-            if Configuration.default_config_file_path() != config.file_path:
+            if default_paths.config_file_path != config.file_path:
                 raise UpgradeFailure('The agent is not using the default configuration file so you may not use the '
                                      'upgrade tarball command.')
-            if Configuration.default_agent_data_path() != config.agent_data_path:
+            if default_paths.agent_data_path != config.agent_data_path:
                 raise UpgradeFailure('The agent is not using the default data directory so you may not use the upgrade '
                                      'tarball command.')
-            if Configuration.default_agent_log_path() != config.agent_log_path:
+            if default_paths.agent_log_path != config.agent_log_path:
                 raise UpgradeFailure('The agent is not using the default log directory so you may not use the upgrade '
                                      'tarball command.')
 
@@ -469,14 +477,16 @@ if __name__ == '__main__':
         print >> sys.stderr, 'You must be root to update the user account that is used to run the agent.'
         sys.exit(1)
 
+    default_paths = PlatformController.new_platform().default_paths
+
     if options.config_filename is None:
-        options.config_filename = Configuration.default_config_file_path()
+        options.config_filename = default_paths.config_file_path
 
     if not os.path.isabs(options.config_filename):
         options.config_filename = os.path.abspath(options.config_filename)
 
     try:
-        config_file = Configuration(options.config_filename, None, None)
+        config_file = Configuration(options.config_filename, default_paths, None, None)
         config_file.parse()
     except Exception, e:
         print >> sys.stderr, 'Error reading configuration file: %s' % str(e)