+++ scalyr_agent/tests/configuration_test.py
@@ -14,6 +14,7 @@
 # ------------------------------------------------------------------------
 #
 # author: Steven Czerwinski <czerwin@scalyr.com>
+from scalyr_agent.platform_controller import DefaultPaths
 
 __author__ = 'czerwin@scalyr.com'
 
@@ -588,4 +589,6 @@ class TestConfiguration(unittest.TestCase):
         def monitor_factory(config, _):
             return TestConfiguration.MonitorObject(config)
 
-        return Configuration(self.__config_file, log_factory, monitor_factory)
+        default_paths = DefaultPaths('/var/log/scalyr-agent-2', '/etc/scalyr-agent-2/agent.json',
+                                     '/var/lib/scalyr-agent-2')
+        return Configuration(self.__config_file, default_paths, log_factory, monitor_factory)